fix configNode code
