+++ confignode/src/main/java/org/apache/iotdb/confignode/consensus/request/read/template/GetPathsSetTemplatePlan.java
@@ -17,61 +17,41 @@
  * under the License.
  */
 
-package org.apache.iotdb.confignode.consensus.request.read;
+package org.apache.iotdb.confignode.consensus.request.read.template;
 
 import org.apache.iotdb.confignode.consensus.request.ConfigPhysicalPlan;
 import org.apache.iotdb.confignode.consensus.request.ConfigPhysicalPlanType;
-import org.apache.iotdb.tsfile.common.conf.TSFileConfig;
 import org.apache.iotdb.tsfile.utils.ReadWriteIOUtils;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.Objects;
 
-public class GetSchemaTemplatePlan extends ConfigPhysicalPlan {
+public class GetPathsSetTemplatePlan extends ConfigPhysicalPlan {
 
-  private String templateName;
+  private String name;
 
-  public GetSchemaTemplatePlan() {
-    super(ConfigPhysicalPlanType.GetSchemaTemplate);
+  public GetPathsSetTemplatePlan() {
+    super(ConfigPhysicalPlanType.GetPathsSetTemplate);
   }
 
-  public GetSchemaTemplatePlan(String templateName) {
-    this();
-    this.templateName = templateName;
+  public GetPathsSetTemplatePlan(String name) {
+    super(ConfigPhysicalPlanType.GetPathsSetTemplate);
+    this.name = name;
   }
 
-  public String getTemplateName() {
-    return templateName;
+  public String getName() {
+    return name;
   }
 
   @Override
   protected void serializeImpl(DataOutputStream stream) throws IOException {
-    stream.writeInt(ConfigPhysicalPlanType.GetSchemaTemplate.ordinal());
-    byte[] bytes = this.getTemplateName().getBytes();
-    int length = bytes.length;
-    stream.writeInt(length);
-    stream.write(bytes);
+    stream.writeInt(ConfigPhysicalPlanType.GetPathsSetTemplate.ordinal());
+    ReadWriteIOUtils.write(name, stream);
   }
 
   @Override
   protected void deserializeImpl(ByteBuffer buffer) throws IOException {
-    int length = ReadWriteIOUtils.readInt(buffer);
-    byte[] dataBytes = ReadWriteIOUtils.readBytes(buffer, length);
-    this.templateName = new String(dataBytes, TSFileConfig.STRING_CHARSET);
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    GetSchemaTemplatePlan that = (GetSchemaTemplatePlan) o;
-    return this.templateName.equalsIgnoreCase(this.templateName);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(templateName);
+    this.name = ReadWriteIOUtils.readString(buffer);
   }
 }
